test-client/requirements.txt
bidict==0.22.1
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
python-engineio==4.4.0
python-socketio==5.8.0
requests==2.28.2
urllib3==1.26.15
websocket-client==1.5.1
